get and display my posts
