+++ frontend/pages/post/friends.js
@@ -1,0 +1,15 @@
+import React from 'react'
+import TopNavigation from '../TopNavigation'
+import SideNav from '../../components/SideNav'
+
+export default function Friends() {
+  return (
+    <div>
+      <TopNavigation />
+      <SideNav />
+      <div className='mt-20 ml-40'>
+        
+      </div>
+    </div>
+  )
+}
